+++ tests/test_annotations.py
@@ -34,6 +34,7 @@
 
 import enum
 import inspect
+import sys
 import typing
 from collections import abc as collections
 from unittest import mock
@@ -3797,3 +3798,161 @@ def test_for_user_option():
     assert tracked_option.key == "arg_2"
     assert tracked_option.name == "arg_2"
     assert tracked_option.type is hikari.OptionType.USER
+
+
+def test_when_annotated_not_top_level():
+    @annotations.with_annotated_args(follow_wrapped=True)
+    @tanjun.as_slash_command("name", "description")
+    @tanjun.as_message_command("name")
+    async def command(
+        ctx: tanjun.abc.Context,
+        *,
+        value: typing.Union[typing.Annotated[annotations.Positional[annotations.Str], "nyaa"], bool] = False,
+        other_value: typing.Optional[typing.Annotated[annotations.Ranged[123, 432], "meow"]] = None,
+    ) -> None:
+        raise NotImplementedError
+
+    assert command.build().options == [
+        hikari.CommandOption(
+            type=hikari.OptionType.STRING,
+            name="value",
+            description="nyaa",
+            is_required=False,
+        ),
+        hikari.CommandOption(
+            type=hikari.OptionType.INTEGER,
+            name="other_value",
+            description="meow",
+            is_required=False,
+            min_value=123,
+            max_value=432,
+        ),
+    ]
+
+    assert len(command._tracked_options) == 2
+    tracked_option = command._tracked_options["value"]
+    assert tracked_option.converters == []
+    assert tracked_option.default is False
+    assert tracked_option.is_always_float is False
+    assert tracked_option.is_only_member is False
+    assert tracked_option.key == "value"
+    assert tracked_option.name == "value"
+    assert tracked_option.type is hikari.OptionType.STRING
+
+    tracked_option = command._tracked_options["other_value"]
+    assert tracked_option.converters == []
+    assert tracked_option.default is None
+    assert tracked_option.is_always_float is False
+    assert tracked_option.is_only_member is False
+    assert tracked_option.key == "other_value"
+    assert tracked_option.name == "other_value"
+    assert tracked_option.type is hikari.OptionType.INTEGER
+
+    assert isinstance(command.wrapped_command, tanjun.MessageCommand)
+    assert isinstance(command.wrapped_command.parser, tanjun.ShlexParser)
+
+    assert len(command.wrapped_command.parser.arguments) == 1
+    argument = command.wrapped_command.parser.arguments[0]
+    assert argument.key == "value"
+    assert argument.converters == []
+    assert argument.default is False
+    assert argument.is_greedy is False
+    assert argument.is_multi is False
+    assert argument.min_value is None
+    assert argument.max_value is None
+
+    assert len(command.wrapped_command.parser.options) == 1
+    option = command.wrapped_command.parser.options[0]
+    assert option.key == "other_value"
+    assert option.names == ["--other-value"]
+    assert option.converters == [int]
+    assert option.default is None
+    assert option.empty_value is tanjun.parsing.UNDEFINED
+    assert option.is_multi is False
+    assert option.min_value == 123
+    assert option.max_value == 432
+
+
+if sys.version_info >= (3, 10):
+
+    def test_when_annotated_not_top_level_3_10_union():
+        @annotations.with_annotated_args(follow_wrapped=True)
+        @tanjun.as_slash_command("name", "description")
+        @tanjun.as_message_command("name")
+        async def command(
+            ctx: tanjun.abc.Context,
+            *,
+            value: typing.Annotated[annotations.Positional[annotations.Str], "nyaa"] | bool = False,
+            other_value: typing.Annotated[annotations.Ranged[123, 432], "meow"] | None = None,
+        ) -> None:
+            raise NotImplementedError
+
+        assert command.build().options == [
+            hikari.CommandOption(
+                type=hikari.OptionType.STRING,
+                name="value",
+                description="nyaa",
+                is_required=False,
+            ),
+            hikari.CommandOption(
+                type=hikari.OptionType.INTEGER,
+                name="other_value",
+                description="meow",
+                is_required=False,
+                min_value=123,
+                max_value=432,
+            ),
+        ]
+
+        assert len(command._tracked_options) == 2
+        tracked_option = command._tracked_options["value"]
+        assert tracked_option.converters == []
+        assert tracked_option.default is False
+        assert tracked_option.is_always_float is False
+        assert tracked_option.is_only_member is False
+        assert tracked_option.key == "value"
+        assert tracked_option.name == "value"
+        assert tracked_option.type is hikari.OptionType.STRING
+
+        tracked_option = command._tracked_options["other_value"]
+        assert tracked_option.converters == []
+        assert tracked_option.default is None
+        assert tracked_option.is_always_float is False
+        assert tracked_option.is_only_member is False
+        assert tracked_option.key == "other_value"
+        assert tracked_option.name == "other_value"
+        assert tracked_option.type is hikari.OptionType.INTEGER
+
+        assert isinstance(command.wrapped_command, tanjun.MessageCommand)
+        assert isinstance(command.wrapped_command.parser, tanjun.ShlexParser)
+
+        assert len(command.wrapped_command.parser.arguments) == 1
+        argument = command.wrapped_command.parser.arguments[0]
+        assert argument.key == "value"
+        assert argument.converters == []
+        assert argument.default is False
+        assert argument.is_greedy is False
+        assert argument.is_multi is False
+        assert argument.min_value is None
+        assert argument.max_value is None
+
+        assert len(command.wrapped_command.parser.options) == 1
+        option = command.wrapped_command.parser.options[0]
+        assert option.key == "other_value"
+        assert option.names == ["--other-value"]
+        assert option.converters == [int]
+        assert option.default is None
+        assert option.empty_value is tanjun.parsing.UNDEFINED
+        assert option.is_multi is False
+        assert option.min_value == 123
+        assert option.max_value == 432
+
+
+def test_when_annotated_handles_unions():
+    ...
+
+
+if sys.version_info >= (3, 10):
+
+    def test_when_annotated_handles_3_10_unions():
+        ...